Move `Rect` to `util/`
