+++ app/gui2/src/components/GraphEditor.vue
@@ -8,11 +8,11 @@ import SelectionBrush from '@/components/SelectionBrush.vue'
 import TopBar from '@/components/TopBar.vue'
 import { useGraphStore } from '@/stores/graph'
 import { useProjectStore } from '@/stores/project'
-import type { Rect } from '@/stores/rect'
 import { useSuggestionDbStore } from '@/stores/suggestionDatabase'
 import { colorFromString } from '@/util/colors'
 import { keyboardBusy, keyboardBusyExceptIn, useEvent, usePointer } from '@/util/events'
 import { useNavigator } from '@/util/navigator'
+import type { Rect } from '@/util/rect'
 import { Vec2 } from '@/util/vec2'
 import * as set from 'lib0/set'
 import type { ContentRange, ExprId } from 'shared/yjsModel'
