+++ app/gui2/src/components/NodeSpan.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { spanKindName, type Span } from '@/stores/graph'
-import { Rect } from '@/stores/rect'
 import { useResizeObserver } from '@/util/events'
+import { Rect } from '@/util/rect'
 import { Vec2 } from '@/util/vec2'
 import type { ExprId } from 'shared/yjsModel'
 import { computed, onUpdated, ref, shallowRef, watch } from 'vue'
