+++ app/gui2/src/util/rect.ts
